refactor(dataset selector): extract get distincts from value selector
