+++ src/ui/units/dash/containers/Dialogs/Control2/Sections/CommonSettingsSection/DatasetSettings/DatasetSettings.tsx
@@ -5,6 +5,9 @@ import {useSelector} from 'react-redux';
 import {DATASET_FIELD_TYPES, DashTabItemControlSourceType, DatasetFieldType} from 'shared';
 
 import {SectionWrapper} from '../../../../../../../../components/SectionWrapper/SectionWrapper';
+import {VIEW_MODES} from '../../../../../../../../components/Select/hooks/useSelectRenderFilter/useSelectRenderFilter';
+import {SelectFeaturedAsyncProps} from '../../../../../../../../components/Select/wrappers/SelectFeaturedAsync';
+import {selectWorkbookId} from '../../../../../../../workbooks/store/selectors';
 import {
     selectSelectorControlType,
     selectSelectorDialog,
@@ -12,9 +15,15 @@ import {
 import {ELEMENT_TYPE} from '../../../../Control/constants';
 import {OperationSelector} from '../../OperationSelector/OperationSelector';
 import {ValueSelector} from '../../ValueSelector/ValueSelector';
+import type {ValueSelectorControlProps} from '../../ValueSelector/types';
 import {DatasetSelector} from '../DatasetSelector/DatasetSelector';
 import {InputTypeSelector} from '../InputTypeSelector/InputTypeSelector';
 import {getElementOptions} from '../helpers/input-type-select';
+
+import {
+    DEFAULT_PAGE_SIZE,
+    getDistinctsByDatasetField,
+} from './helpers/get-distincts-by-dataset-field';
 
 const i18n = I18n.keyset('dash.control-dialog.edit');
 
@@ -23,8 +32,13 @@ type DatasetSettingsProps = {
 };
 
 const DatasetSettings = ({isSectionHidden}: DatasetSettingsProps) => {
-    const {datasetFieldType, sourceType, fieldType} = useSelector(selectSelectorDialog);
+    const {datasetFieldType, datasetId, datasetFieldId, sourceType, fieldType} =
+        useSelector(selectSelectorDialog);
     const controlType = useSelector(selectSelectorControlType);
+    const workbookId = useSelector(selectWorkbookId);
+
+    const [searchPattern, setSearchPattern] = React.useState('');
+
     const options = React.useMemo(() => {
         const disabledOptions: Record<string, boolean> = {
             [ELEMENT_TYPE.DATE]:
@@ -44,6 +58,36 @@ const DatasetSettings = ({isSectionHidden}: DatasetSettingsProps) => {
         }));
     }, [sourceType, fieldType, controlType, datasetFieldType]);
 
+    const onFilterChange = React.useCallback<
+        NonNullable<SelectFeaturedAsyncProps['onFilterChange']>
+    >((pattern, mode) => {
+        if (mode === VIEW_MODES.ALL) {
+            setSearchPattern(pattern);
+        }
+    }, []);
+
+    const fetcher = React.useCallback(
+        ({pageNumber, pageSize} = {pageNumber: 0, pageSize: DEFAULT_PAGE_SIZE}) =>
+            getDistinctsByDatasetField({
+                datasetId,
+                workbookId,
+                datasetFieldId,
+                nextPageToken: pageNumber,
+                searchPattern,
+                pageSize,
+            }),
+        [datasetId, workbookId, datasetFieldId, searchPattern],
+    );
+
+    const controlProps: ValueSelectorControlProps = React.useMemo(() => {
+        return {
+            select: {
+                type: 'dynamic',
+                custom: {fetcher, onFilterChange, disabled: !datasetId || !datasetFieldId},
+            },
+        };
+    }, [datasetFieldId, datasetId, fetcher, onFilterChange]);
+
     return (
         <React.Fragment>
             <SectionWrapper
@@ -53,7 +97,7 @@ const DatasetSettings = ({isSectionHidden}: DatasetSettingsProps) => {
                 <DatasetSelector />
                 <InputTypeSelector options={options} />
                 <OperationSelector />
-                <ValueSelector />
+                <ValueSelector controlProps={controlProps} />
             </SectionWrapper>
         </React.Fragment>
     );
